Update activity level options and property types
Changed activity level options in `AdditionalInfo.cshtml` to use commas as decimal separators. Updated `Height` property type from `float` to `int` and `ActivityLevel` from `float` to `decimal` in `AdditionalInfo.cshtml.cs` for improved data accuracy.

+++ Fitness/Pages/Login/AdditionalInfo.cshtml.cs
@@ -17,7 +17,7 @@ namespace Fitness.Pages.Login
         public decimal Weight { get; set; }
 
         [BindProperty]
-        public float Height { get; set; }
+        public int Height { get; set; }
 
         [BindProperty]
         public int Age { get; set; }
@@ -26,7 +26,7 @@ namespace Fitness.Pages.Login
         public string Gender { get; set; }
 
         [BindProperty]
-        public float ActivityLevel { get; set; }
+        public decimal ActivityLevel { get; set; }
 
         public IActionResult OnPost()
         {
